requirements.txt
affine==2.4.0
annotated-types==0.6.0
anyio==4.2.0
astroid==3.0.2
attrs==23.2.0
blinker==1.7.0
boto3==1.23.10
botocore==1.26.10
certifi==2023.11.17
charset-normalizer==2.0.12
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
deprecation==2.1.0
dill==0.3.7
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
geojson==2.5.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==3.0.1
jsonseq==1.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mercantile==1.1.6
numpy==1.26.3
packaging==23.2
platformdirs==4.1.0
pluggy==1.4.0
postgrest==0.15.0
pydantic==2.6.1
pydantic_core==2.16.2
pylint==3.0.3
pyparsing==3.1.1
pyrsistent==0.20.0
pytest==8.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
rasterio==1.3.9
requests==2.27.1
requests-toolbelt==0.9.1
s3transfer==0.5.2
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
SQLAlchemy==2.0.30
StrEnum==0.4.15
supermercado==0.2.0
tomlkit==0.12.3
typing_extensions==4.9.0
urllib3==1.26.18
Werkzeug==3.0.1
